pyproject.toml
[project]
name = "chatgpt"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "enginearn", email = "enginyaaaaan@gmail.com" }
]
dependencies = [
    "django-bootstrap5>=23.3",
    "openai>=0.27.8",
    "langchain>=0.0.27",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["black~=23.3.0", "mypy~=1.3.0", "isort~=5.12.0", "django~=4.2.1", "openai~=0.27.7"]
[tool.hatch.metadata]
allow-direct-references = true
